+++ pygal/graph/bar.py
@@ -91,7 +91,7 @@ class Bar(Graph):
             self._tooltip_data(
                 bar, val, x_center, y_center, "centered",
                 self._get_x_label(i))
-            self._static_value(serie_node, val, x_center, y_center)
+            self._static_value(serie_node, val, x_center, y_center, metadata)
 
     def _compute(self):
         """Compute y min and max and y scale and set labels"""