+++ pygal/graph/pie.py
@@ -84,7 +84,7 @@ class Pie(Graph):
 
             alter(self.svg.slice(
                 serie_node, slice_, big_radius, small_radius,
-                angle, start_angle, center, val, i), metadata)
+                angle, start_angle, center, val, i, metadata), metadata)
             start_angle += angle
             total_perc += perc
 
@@ -93,7 +93,7 @@ class Pie(Graph):
             self.svg.slice(serie_node,
                            self.svg.node(slices, class_="big_slice"),
                            radius * .9, 0, serie_angle,
-                           original_start_angle, center, val, i)
+                           original_start_angle, center, val, i, metadata)
         return serie_angle
 
     def _compute_x_labels(self):